Add files via upload
V0.2. COMBAT BETA

For the sake of actually making a working game, I have chosen to disregard the majority of the classes in favor of a simple turn-based system in the GameDriver class. As it stands, you can Attack or Defend in combat, and can challenge one of three types of slime enemies: Verdant (easy), Mud (medium), or Ice (hard). They all handle relatively the same but with harder ones having beefier stats. The game ends after slaying a slime, but can be restarted or quit ad nauseum. 

KNOWN BUGS:
-The enemy still gets in its last move after its HP drops to or below 0. This is purely a visual issue and is remedied between battles.
-Enemy HP can go into the negatives. Again, this is purely a visual issue. 

-Keillor C.